Revealable text fields, both expanding and masking
Certain fields in the details pane must be ellipsized and then
expand themselves when clicked on. The ellipsizing should be
driven by the width of the field, not the number of characters.

Certain fields in the details pane should be masked as a courtesy,
such as password environment variables. Note that this does not
substitute for real security such as use of kubernetes secrets,
instead of environment variables.

Both ellipsized and masked fields are revealed in full when clicked on.

This patch includes an implementation of this behavior.

Some parts of this behavior are CSS driven, and an example CSS
file is included.

Fixes #16
Closes #21